requirements.txt
pandas
numpy
scikit-learn
rouge-score
tqdm
evaluate
bert-score
konlpy
nltk
wandb
python-dotenv
unsloth_zoo
vllm
unsloth
xformers
accelerate
trl
peft
transformers==4.53.2